Put the openai client in a separate component
