+++ cloud-tenant-base-dependencies-enforcer/pom.xml
@@ -235,18 +235,6 @@
                                         <include>org.osgi:org.osgi.compendium:4.1.0:test</include>
                                         <include>org.osgi:org.osgi.core:4.1.0:test</include>
                                         <include>xerces:xercesImpl:2.12.2:test</include>
-
-                                        <include>com.squareup.okhttp3:okhttp:3.14.9:test</include>
-                                        <include>com.squareup.okio:okio:1.17.2:test</include>
-                                        <include>com.squareup.retrofit2:adapter-rxjava2:2.9.0:test</include>
-                                        <include>com.squareup.retrofit2:converter-jackson:2.9.0:test</include>
-                                        <include>com.squareup.retrofit2:retrofit:2.9.0:test</include>
-                                        <include>com.theokanning.openai-gpt3-java:api:${openai-gpt3.version}:test</include>
-                                        <include>com.theokanning.openai-gpt3-java:client:${openai-gpt3.version}:test</include>
-                                        <include>com.theokanning.openai-gpt3-java:service:${openai-gpt3.version}:test</include>
-                                        <include>io.reactivex.rxjava2:rxjava:2.0.0:test</include>
-                                        <include>org.reactivestreams:reactive-streams:1.0.3:test</include>
-
                                     </allowed>
                                 </enforceDependencies>
                             </rules>
