+++ model-integration/src/main/java/ai/vespa/llm/generation/GeneratorOptions.java
@@ -1,4 +1,4 @@
-package ai.vespa.llm;
+package ai.vespa.llm.generation;
 
 import com.yahoo.api.annotations.Beta;
