+++ vespajlib/src/main/java/ai/vespa/llm/completion/package-info.java
@@ -1,11 +1,11 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 @ExportPackage
 @PublicApi
-package ai.vespa.llm.client;
+package ai.vespa.llm.completion;
 
 import com.yahoo.api.annotations.PublicApi;
 import com.yahoo.osgi.annotation.ExportPackage;
 
 /**
- * Clients to externally hosted large language models.
+ * Classes for generating text completions with language models.
  */
