+++ vespajlib/src/test/java/ai/vespa/llm/completion/CompletionTest.java
@@ -1,5 +1,8 @@
-package ai.vespa.llm;
+package ai.vespa.llm.completion;
 
+import ai.vespa.llm.completion.Completion;
+import ai.vespa.llm.completion.Prompt;
+import ai.vespa.llm.completion.StringPrompt;
 import ai.vespa.llm.test.MockLanguageModel;
 import org.junit.jupiter.api.Test;
